feat: add ShareUtil for app sharing functionality
