+++ feature/webview/src/main/java/com/nexters/misik/webview/WebViewScreen.kt
@@ -4,7 +4,6 @@ import android.annotation.SuppressLint
 import android.content.ClipData
 import android.content.ClipboardManager
 import android.content.Context
-import android.content.Intent
 import androidx.compose.foundation.layout.Box
 import androidx.compose.foundation.layout.fillMaxSize
 import androidx.compose.runtime.Composable
@@ -17,11 +16,11 @@ import androidx.compose.ui.platform.LocalContext
 import androidx.compose.ui.viewinterop.AndroidView
 import androidx.hilt.navigation.compose.hiltViewModel
 import androidx.lifecycle.compose.collectAsStateWithLifecycle
-import com.nexters.misik.feature.webview.R
 import com.nexters.misik.preview.PreviewService
 import com.nexters.misik.webview.base.MisikWebViewFactory
 import com.nexters.misik.webview.bridge.WebInterface
 import com.nexters.misik.webview.common.LoadingAnimation
+import com.nexters.misik.webview.util.ShareUtil
 import timber.log.Timber
 
 @SuppressLint("SetJavaScriptEnabled", "JavascriptInterface")
@@ -51,7 +50,7 @@ fun WebViewScreen(
                 )
 
                 is WebViewIntent.Share -> {
-                    shareText(context)
+                    ShareUtil.shareApp(context)
                 }
 
                 else -> viewModel.sendIntent(intent)
@@ -102,16 +101,6 @@ fun WebViewScreen(
             }
         }
     }
-}
-
-private fun shareText(context: Context) {
-    val title = context.getString(R.string.share_title)
-    val description = context.getString(R.string.share_description)
-    val shareIntent = Intent(Intent.ACTION_SEND).apply {
-        type = "text/plain"
-        putExtra(Intent.EXTRA_TEXT, description)
-    }
-    context.startActivity(Intent.createChooser(shareIntent, title))
 }
 
 @Composable
